Refs #19794: Added OMG idl for rpc types
