+++ src/cpp/fastdds/builtin/type_lookup_service/detail/TypeLookupTypesPubSubTypes.h
@@ -31,8 +31,9 @@
 
 #include "TypeLookupTypes.hpp"
 
-#include <fastdds/dds/xtypes/type_representation/TypeObject.hpp>
+#include "rpc_typesPubSubTypes.h"
 #include <fastdds/dds/core/ReturnCode.hpp>
+#include <fastdds/dds/xtypes/type_representation/TypeObject.hpp>
 
 #if !defined(GEN_API_VER) || (GEN_API_VER != 2)
 #error \
@@ -45,555 +46,6 @@ namespace eprosima
     {
         namespace dds
         {
-            typedef std::array<uint8_t, 12> GuidPrefix_t;
-
-
-
-            /*!
-             * @brief This class represents the TopicDataType of the type EntityId_t defined by the user in the IDL file.
-             * @ingroup TypeLookupTypes
-             */
-            class EntityId_tPubSubType : public eprosima::fastdds::dds::TopicDataType
-            {
-            public:
-
-                typedef EntityId_t type;
-
-                eProsima_user_DllExport EntityId_tPubSubType();
-
-                eProsima_user_DllExport ~EntityId_tPubSubType() override;
-
-                eProsima_user_DllExport bool serialize(
-                        void* data,
-                        eprosima::fastrtps::rtps::SerializedPayload_t* payload) override
-                {
-                    return serialize(data, payload, eprosima::fastdds::dds::DEFAULT_DATA_REPRESENTATION);
-                }
-
-                eProsima_user_DllExport bool serialize(
-                        void* data,
-                        eprosima::fastrtps::rtps::SerializedPayload_t* payload,
-                        eprosima::fastdds::dds::DataRepresentationId_t data_representation) override;
-
-                eProsima_user_DllExport bool deserialize(
-                        eprosima::fastrtps::rtps::SerializedPayload_t* payload,
-                        void* data) override;
-
-                eProsima_user_DllExport std::function<uint32_t()> getSerializedSizeProvider(
-                        void* data) override
-                {
-                    return getSerializedSizeProvider(data, eprosima::fastdds::dds::DEFAULT_DATA_REPRESENTATION);
-                }
-
-                eProsima_user_DllExport std::function<uint32_t()> getSerializedSizeProvider(
-                        void* data,
-                        eprosima::fastdds::dds::DataRepresentationId_t data_representation) override;
-
-                eProsima_user_DllExport bool getKey(
-                        void* data,
-                        eprosima::fastrtps::rtps::InstanceHandle_t* ihandle,
-                        bool force_md5 = false) override;
-
-                eProsima_user_DllExport void* createData() override;
-
-                eProsima_user_DllExport void deleteData(
-                        void* data) override;
-
-            #ifdef TOPIC_DATA_TYPE_API_HAS_IS_BOUNDED
-                eProsima_user_DllExport inline bool is_bounded() const override
-                {
-                    return true;
-                }
-
-            #endif  // TOPIC_DATA_TYPE_API_HAS_IS_BOUNDED
-
-            #ifdef TOPIC_DATA_TYPE_API_HAS_IS_PLAIN
-                eProsima_user_DllExport inline bool is_plain() const override
-                {
-                    return false;
-                }
-
-                eProsima_user_DllExport inline bool is_plain(
-                    eprosima::fastdds::dds::DataRepresentationId_t data_representation) const override
-                {
-                    static_cast<void>(data_representation);
-                    return false;
-                }
-
-            #endif  // TOPIC_DATA_TYPE_API_HAS_IS_PLAIN
-
-            #ifdef TOPIC_DATA_TYPE_API_HAS_CONSTRUCT_SAMPLE
-                eProsima_user_DllExport inline bool construct_sample(
-                        void* memory) const override
-                {
-                    static_cast<void>(memory);
-                    return false;
-                }
-
-            #endif  // TOPIC_DATA_TYPE_API_HAS_CONSTRUCT_SAMPLE
-
-                MD5 m_md5;
-                unsigned char* m_keyBuffer;
-
-            };
-
-
-
-            /*!
-             * @brief This class represents the TopicDataType of the type GUID_t defined by the user in the IDL file.
-             * @ingroup TypeLookupTypes
-             */
-            class GUID_tPubSubType : public eprosima::fastdds::dds::TopicDataType
-            {
-            public:
-
-                typedef GUID_t type;
-
-                eProsima_user_DllExport GUID_tPubSubType();
-
-                eProsima_user_DllExport ~GUID_tPubSubType() override;
-
-                eProsima_user_DllExport bool serialize(
-                        void* data,
-                        eprosima::fastrtps::rtps::SerializedPayload_t* payload) override
-                {
-                    return serialize(data, payload, eprosima::fastdds::dds::DEFAULT_DATA_REPRESENTATION);
-                }
-
-                eProsima_user_DllExport bool serialize(
-                        void* data,
-                        eprosima::fastrtps::rtps::SerializedPayload_t* payload,
-                        eprosima::fastdds::dds::DataRepresentationId_t data_representation) override;
-
-                eProsima_user_DllExport bool deserialize(
-                        eprosima::fastrtps::rtps::SerializedPayload_t* payload,
-                        void* data) override;
-
-                eProsima_user_DllExport std::function<uint32_t()> getSerializedSizeProvider(
-                        void* data) override
-                {
-                    return getSerializedSizeProvider(data, eprosima::fastdds::dds::DEFAULT_DATA_REPRESENTATION);
-                }
-
-                eProsima_user_DllExport std::function<uint32_t()> getSerializedSizeProvider(
-                        void* data,
-                        eprosima::fastdds::dds::DataRepresentationId_t data_representation) override;
-
-                eProsima_user_DllExport bool getKey(
-                        void* data,
-                        eprosima::fastrtps::rtps::InstanceHandle_t* ihandle,
-                        bool force_md5 = false) override;
-
-                eProsima_user_DllExport void* createData() override;
-
-                eProsima_user_DllExport void deleteData(
-                        void* data) override;
-
-            #ifdef TOPIC_DATA_TYPE_API_HAS_IS_BOUNDED
-                eProsima_user_DllExport inline bool is_bounded() const override
-                {
-                    return true;
-                }
-
-            #endif  // TOPIC_DATA_TYPE_API_HAS_IS_BOUNDED
-
-            #ifdef TOPIC_DATA_TYPE_API_HAS_IS_PLAIN
-                eProsima_user_DllExport inline bool is_plain() const override
-                {
-                    return false;
-                }
-
-                eProsima_user_DllExport inline bool is_plain(
-                    eprosima::fastdds::dds::DataRepresentationId_t data_representation) const override
-                {
-                    static_cast<void>(data_representation);
-                    return false;
-                }
-
-            #endif  // TOPIC_DATA_TYPE_API_HAS_IS_PLAIN
-
-            #ifdef TOPIC_DATA_TYPE_API_HAS_CONSTRUCT_SAMPLE
-                eProsima_user_DllExport inline bool construct_sample(
-                        void* memory) const override
-                {
-                    static_cast<void>(memory);
-                    return false;
-                }
-
-            #endif  // TOPIC_DATA_TYPE_API_HAS_CONSTRUCT_SAMPLE
-
-                MD5 m_md5;
-                unsigned char* m_keyBuffer;
-
-            };
-
-
-
-            /*!
-             * @brief This class represents the TopicDataType of the type SequenceNumber_t defined by the user in the IDL file.
-             * @ingroup TypeLookupTypes
-             */
-            class SequenceNumber_tPubSubType : public eprosima::fastdds::dds::TopicDataType
-            {
-            public:
-
-                typedef SequenceNumber_t type;
-
-                eProsima_user_DllExport SequenceNumber_tPubSubType();
-
-                eProsima_user_DllExport ~SequenceNumber_tPubSubType() override;
-
-                eProsima_user_DllExport bool serialize(
-                        void* data,
-                        eprosima::fastrtps::rtps::SerializedPayload_t* payload) override
-                {
-                    return serialize(data, payload, eprosima::fastdds::dds::DEFAULT_DATA_REPRESENTATION);
-                }
-
-                eProsima_user_DllExport bool serialize(
-                        void* data,
-                        eprosima::fastrtps::rtps::SerializedPayload_t* payload,
-                        eprosima::fastdds::dds::DataRepresentationId_t data_representation) override;
-
-                eProsima_user_DllExport bool deserialize(
-                        eprosima::fastrtps::rtps::SerializedPayload_t* payload,
-                        void* data) override;
-
-                eProsima_user_DllExport std::function<uint32_t()> getSerializedSizeProvider(
-                        void* data) override
-                {
-                    return getSerializedSizeProvider(data, eprosima::fastdds::dds::DEFAULT_DATA_REPRESENTATION);
-                }
-
-                eProsima_user_DllExport std::function<uint32_t()> getSerializedSizeProvider(
-                        void* data,
-                        eprosima::fastdds::dds::DataRepresentationId_t data_representation) override;
-
-                eProsima_user_DllExport bool getKey(
-                        void* data,
-                        eprosima::fastrtps::rtps::InstanceHandle_t* ihandle,
-                        bool force_md5 = false) override;
-
-                eProsima_user_DllExport void* createData() override;
-
-                eProsima_user_DllExport void deleteData(
-                        void* data) override;
-
-            #ifdef TOPIC_DATA_TYPE_API_HAS_IS_BOUNDED
-                eProsima_user_DllExport inline bool is_bounded() const override
-                {
-                    return true;
-                }
-
-            #endif  // TOPIC_DATA_TYPE_API_HAS_IS_BOUNDED
-
-            #ifdef TOPIC_DATA_TYPE_API_HAS_IS_PLAIN
-                eProsima_user_DllExport inline bool is_plain() const override
-                {
-                    return false;
-                }
-
-                eProsima_user_DllExport inline bool is_plain(
-                    eprosima::fastdds::dds::DataRepresentationId_t data_representation) const override
-                {
-                    static_cast<void>(data_representation);
-                    return false;
-                }
-
-            #endif  // TOPIC_DATA_TYPE_API_HAS_IS_PLAIN
-
-            #ifdef TOPIC_DATA_TYPE_API_HAS_CONSTRUCT_SAMPLE
-                eProsima_user_DllExport inline bool construct_sample(
-                        void* memory) const override
-                {
-                    static_cast<void>(memory);
-                    return false;
-                }
-
-            #endif  // TOPIC_DATA_TYPE_API_HAS_CONSTRUCT_SAMPLE
-
-                MD5 m_md5;
-                unsigned char* m_keyBuffer;
-
-            };
-
-
-
-            /*!
-             * @brief This class represents the TopicDataType of the type SampleIdentity defined by the user in the IDL file.
-             * @ingroup TypeLookupTypes
-             */
-            class SampleIdentityPubSubType : public eprosima::fastdds::dds::TopicDataType
-            {
-            public:
-
-                typedef SampleIdentity type;
-
-                eProsima_user_DllExport SampleIdentityPubSubType();
-
-                eProsima_user_DllExport ~SampleIdentityPubSubType() override;
-
-                eProsima_user_DllExport bool serialize(
-                        void* data,
-                        eprosima::fastrtps::rtps::SerializedPayload_t* payload) override
-                {
-                    return serialize(data, payload, eprosima::fastdds::dds::DEFAULT_DATA_REPRESENTATION);
-                }
-
-                eProsima_user_DllExport bool serialize(
-                        void* data,
-                        eprosima::fastrtps::rtps::SerializedPayload_t* payload,
-                        eprosima::fastdds::dds::DataRepresentationId_t data_representation) override;
-
-                eProsima_user_DllExport bool deserialize(
-                        eprosima::fastrtps::rtps::SerializedPayload_t* payload,
-                        void* data) override;
-
-                eProsima_user_DllExport std::function<uint32_t()> getSerializedSizeProvider(
-                        void* data) override
-                {
-                    return getSerializedSizeProvider(data, eprosima::fastdds::dds::DEFAULT_DATA_REPRESENTATION);
-                }
-
-                eProsima_user_DllExport std::function<uint32_t()> getSerializedSizeProvider(
-                        void* data,
-                        eprosima::fastdds::dds::DataRepresentationId_t data_representation) override;
-
-                eProsima_user_DllExport bool getKey(
-                        void* data,
-                        eprosima::fastrtps::rtps::InstanceHandle_t* ihandle,
-                        bool force_md5 = false) override;
-
-                eProsima_user_DllExport void* createData() override;
-
-                eProsima_user_DllExport void deleteData(
-                        void* data) override;
-
-            #ifdef TOPIC_DATA_TYPE_API_HAS_IS_BOUNDED
-                eProsima_user_DllExport inline bool is_bounded() const override
-                {
-                    return true;
-                }
-
-            #endif  // TOPIC_DATA_TYPE_API_HAS_IS_BOUNDED
-
-            #ifdef TOPIC_DATA_TYPE_API_HAS_IS_PLAIN
-                eProsima_user_DllExport inline bool is_plain() const override
-                {
-                    return false;
-                }
-
-                eProsima_user_DllExport inline bool is_plain(
-                    eprosima::fastdds::dds::DataRepresentationId_t data_representation) const override
-                {
-                    static_cast<void>(data_representation);
-                    return false;
-                }
-
-            #endif  // TOPIC_DATA_TYPE_API_HAS_IS_PLAIN
-
-            #ifdef TOPIC_DATA_TYPE_API_HAS_CONSTRUCT_SAMPLE
-                eProsima_user_DllExport inline bool construct_sample(
-                        void* memory) const override
-                {
-                    static_cast<void>(memory);
-                    return false;
-                }
-
-            #endif  // TOPIC_DATA_TYPE_API_HAS_CONSTRUCT_SAMPLE
-
-                MD5 m_md5;
-                unsigned char* m_keyBuffer;
-
-            };
-            namespace rpc
-            {
-                typedef uint8_t UnknownOperation;
-                typedef uint8_t UnknownException;
-                typedef uint8_t UnusedMember;
-
-                typedef eprosima::fastcdr::fixed_string<255> InstanceName;
-
-
-
-                /*!
-                 * @brief This class represents the TopicDataType of the type RequestHeader defined by the user in the IDL file.
-                 * @ingroup TypeLookupTypes
-                 */
-                class RequestHeaderPubSubType : public eprosima::fastdds::dds::TopicDataType
-                {
-                public:
-
-                    typedef RequestHeader type;
-
-                    eProsima_user_DllExport RequestHeaderPubSubType();
-
-                    eProsima_user_DllExport ~RequestHeaderPubSubType() override;
-
-                    eProsima_user_DllExport bool serialize(
-                            void* data,
-                            eprosima::fastrtps::rtps::SerializedPayload_t* payload) override
-                    {
-                        return serialize(data, payload, eprosima::fastdds::dds::DEFAULT_DATA_REPRESENTATION);
-                    }
-
-                    eProsima_user_DllExport bool serialize(
-                            void* data,
-                            eprosima::fastrtps::rtps::SerializedPayload_t* payload,
-                            eprosima::fastdds::dds::DataRepresentationId_t data_representation) override;
-
-                    eProsima_user_DllExport bool deserialize(
-                            eprosima::fastrtps::rtps::SerializedPayload_t* payload,
-                            void* data) override;
-
-                    eProsima_user_DllExport std::function<uint32_t()> getSerializedSizeProvider(
-                            void* data) override
-                    {
-                        return getSerializedSizeProvider(data, eprosima::fastdds::dds::DEFAULT_DATA_REPRESENTATION);
-                    }
-
-                    eProsima_user_DllExport std::function<uint32_t()> getSerializedSizeProvider(
-                            void* data,
-                            eprosima::fastdds::dds::DataRepresentationId_t data_representation) override;
-
-                    eProsima_user_DllExport bool getKey(
-                            void* data,
-                            eprosima::fastrtps::rtps::InstanceHandle_t* ihandle,
-                            bool force_md5 = false) override;
-
-                    eProsima_user_DllExport void* createData() override;
-
-                    eProsima_user_DllExport void deleteData(
-                            void* data) override;
-
-                #ifdef TOPIC_DATA_TYPE_API_HAS_IS_BOUNDED
-                    eProsima_user_DllExport inline bool is_bounded() const override
-                    {
-                        return true;
-                    }
-
-                #endif  // TOPIC_DATA_TYPE_API_HAS_IS_BOUNDED
-
-                #ifdef TOPIC_DATA_TYPE_API_HAS_IS_PLAIN
-                    eProsima_user_DllExport inline bool is_plain() const override
-                    {
-                        return false;
-                    }
-
-                    eProsima_user_DllExport inline bool is_plain(
-                        eprosima::fastdds::dds::DataRepresentationId_t data_representation) const override
-                    {
-                        static_cast<void>(data_representation);
-                        return false;
-                    }
-
-                #endif  // TOPIC_DATA_TYPE_API_HAS_IS_PLAIN
-
-                #ifdef TOPIC_DATA_TYPE_API_HAS_CONSTRUCT_SAMPLE
-                    eProsima_user_DllExport inline bool construct_sample(
-                            void* memory) const override
-                    {
-                        static_cast<void>(memory);
-                        return false;
-                    }
-
-                #endif  // TOPIC_DATA_TYPE_API_HAS_CONSTRUCT_SAMPLE
-
-                    MD5 m_md5;
-                    unsigned char* m_keyBuffer;
-
-                };
-
-
-
-                /*!
-                 * @brief This class represents the TopicDataType of the type ReplyHeader defined by the user in the IDL file.
-                 * @ingroup TypeLookupTypes
-                 */
-                class ReplyHeaderPubSubType : public eprosima::fastdds::dds::TopicDataType
-                {
-                public:
-
-                    typedef ReplyHeader type;
-
-                    eProsima_user_DllExport ReplyHeaderPubSubType();
-
-                    eProsima_user_DllExport ~ReplyHeaderPubSubType() override;
-
-                    eProsima_user_DllExport bool serialize(
-                            void* data,
-                            eprosima::fastrtps::rtps::SerializedPayload_t* payload) override
-                    {
-                        return serialize(data, payload, eprosima::fastdds::dds::DEFAULT_DATA_REPRESENTATION);
-                    }
-
-                    eProsima_user_DllExport bool serialize(
-                            void* data,
-                            eprosima::fastrtps::rtps::SerializedPayload_t* payload,
-                            eprosima::fastdds::dds::DataRepresentationId_t data_representation) override;
-
-                    eProsima_user_DllExport bool deserialize(
-                            eprosima::fastrtps::rtps::SerializedPayload_t* payload,
-                            void* data) override;
-
-                    eProsima_user_DllExport std::function<uint32_t()> getSerializedSizeProvider(
-                            void* data) override
-                    {
-                        return getSerializedSizeProvider(data, eprosima::fastdds::dds::DEFAULT_DATA_REPRESENTATION);
-                    }
-
-                    eProsima_user_DllExport std::function<uint32_t()> getSerializedSizeProvider(
-                            void* data,
-                            eprosima::fastdds::dds::DataRepresentationId_t data_representation) override;
-
-                    eProsima_user_DllExport bool getKey(
-                            void* data,
-                            eprosima::fastrtps::rtps::InstanceHandle_t* ihandle,
-                            bool force_md5 = false) override;
-
-                    eProsima_user_DllExport void* createData() override;
-
-                    eProsima_user_DllExport void deleteData(
-                            void* data) override;
-
-                #ifdef TOPIC_DATA_TYPE_API_HAS_IS_BOUNDED
-                    eProsima_user_DllExport inline bool is_bounded() const override
-                    {
-                        return false;
-                    }
-
-                #endif  // TOPIC_DATA_TYPE_API_HAS_IS_BOUNDED
-
-                #ifdef TOPIC_DATA_TYPE_API_HAS_IS_PLAIN
-                    eProsima_user_DllExport inline bool is_plain() const override
-                    {
-                        return false;
-                    }
-
-                    eProsima_user_DllExport inline bool is_plain(
-                        eprosima::fastdds::dds::DataRepresentationId_t data_representation) const override
-                    {
-                        static_cast<void>(data_representation);
-                        return false;
-                    }
-
-                #endif  // TOPIC_DATA_TYPE_API_HAS_IS_PLAIN
-
-                #ifdef TOPIC_DATA_TYPE_API_HAS_CONSTRUCT_SAMPLE
-                    eProsima_user_DllExport inline bool construct_sample(
-                            void* memory) const override
-                    {
-                        static_cast<void>(memory);
-                        return false;
-                    }
-
-                #endif  // TOPIC_DATA_TYPE_API_HAS_CONSTRUCT_SAMPLE
-
-                    MD5 m_md5;
-                    unsigned char* m_keyBuffer;
-
-                };
-            } // namespace rpc
             namespace builtin
             {
